+++ NEWS.md
@@ -8,8 +8,8 @@
 
 * Added a new `toolbar()` component for creating Bootstrap toolbars that can contain buttons, text, and other elements. (#1247)
   * Added `toolbar_input_button()` for easily creating buttons to include in a `toolbar()`. (#1248)
-
-* Added a new `toolbar()` component for creating Bootstrap toolbars that can contain buttons, text, and other elements. (#1247)
+  
+  * Added `toolbar_input_select()`, a select input designed for use within a `toolbar()`. (#1249)
 
 ## Improvements and bug fixes